Update Preprocessing_images.py
This code is used for the Deep Learning course at the RuG in order to preprocess images used in the first assignment. The images will be used to train a convolutional neural network based on the VGG-16 architecture to classify MRI images either as containing a tumor or no tumor. The images need to be turned from RGB into grayscale and resized to 224 by 224 pixels.